Separate foreign keys for a reference table by name.
Without this distinction, the code aggregates all the FKs refering to a
given table into a single FK, which is clearly wrong. Left for another
day is the task of adding the name to the constraint created on the
child.

Discovered on version 1.8.7, not sure how far back it needs to apply.