pg_partman_bgw: fix buffer overflow in bgw_name with long database names
The fixed format string that's used takes up 46 bytes and a database name
may be up to 63 bytes long but worker.bgw_name's size is just 64 bytes so
we'd overflow the buffer with a 18 byte database name.

This commit adds overflow checking and truncates long database names.  The
fixed part of the bgw process name could be adjusted to something shorter to
allow for longer database names in the worker name.