pg_partman_bgw: remove a superfluous continue; statement after elog(FATAL)
It appears that the previous code was missing curly braces around an if
block body causing `continue;` to be always executed.  This could've been
fixed by adding the curly braces, but elog(FATAL) causes the process to exit
immediately so there's no point in trying to affect the control flow after
it.