pg_partman_bgw: quote schema names when looking for partman extension
This allows bgw to handle schemas with special characters in them, plus it
avoids a vulnerability in environments where pgextwlist or similar
mechanisms allow non-superusers to install extensions.

Otherwise a user could create a schema called, for example "foo();" and
install the extension there, causing the background worker to call that
function with superuser privileges.