Bump version to 5.0.0-beta
It is useful for update/install files generator three commits later.

The old install file is copied from a state as of tag v4.7.3:
updates/pg_partman--4.7.3.sql

It is useful to have two of them: install file and previous one to check
upgrade script.

Install the newest is usual:
... # CREATE EXTENSION pg_partman;

Test of upgrade can be done via commands:
... # CREATE EXTENSION pg_partman VERSION "4.7.3";
... # ALTER EXTENSION pg_partman UPDATE;